Delete attackable protocol update and proof sketch
We realized that the proposed protocol update was not secure. According
to this approach, an honest miner did not consider adversarially
generated blocks containing false interlinks when constructing the
interlink for the new blocks to be mined. In this way, there were no
interlink pointers from an honestly generated block to any adversarially
generated blocks, thus the later cannot participate in an honest suffix
proof even if they do belong in the honest chain. Now consider the
following attack: the adversary performs chainsewing attack and selfish
mining on the honest players' chain. If n/(n-t) > 1/3 then the adversary
has successfully attacked the 0-level chain quality, in terms of higher
percentage of adversarial versus honest blocks in the chain. Consider
that the adversary has also created a small fork chain and constructs a
suffix proof for that chain including, via chainsewing them, the blocks
that she mined in the honest player's chain. In such a case the suffix
proof constructed by an honest player cannot prevail over the proof of
the adversary.